12/07 신고 중복 체크
 신고 중복 체크
