+++ travelCompass/src/main/java/kh/team/travelcompass/management/model/dao/ManagementDAO.java
@@ -56,17 +56,6 @@ public class ManagementDAO {
 		return sqlSession.update("reportMapper.memberBanCancel", memberNo);
 	}	
 
-	//중복 신고 체크
-	public int reportCheck(Map<String, Object> paramMap) {
-		
-		return sqlSession.selectOne("reportMapper.reportCheck", paramMap);
-	}
-	
-	//신고 등록
-	public int insertReport(Map<String, Object> paramMap) {
-		
-		return sqlSession.insert("reportMapper.insertReport", paramMap);
-	}
