+++ travelCompass/src/main/java/kh/team/travelcompass/report/model/service/ReportService.java
@@ -1,0 +1,16 @@
+package kh.team.travelcompass.report.model.service;
+
+import java.util.Map;
+
+public interface ReportService {
+
+	
+	//중복신고 체크
+	int reportCheck(Map<String, Object> paramMap);	
+	
+	//신고 등록
+	int insertReport(Map<String, Object> paramMap);
+
+	
+	
+}
