+++ travelCompass/src/main/webapp/resources/js/place/review.js
@@ -804,19 +804,24 @@ function insertReport(reviewNo, memberNo, loginMemberNo) {
 
   if (memberNo != loginMemberNo) {
     $.ajax ({
-      url: "/management/insertReport",
+      url: "/insertReport",
       type: "get",
       data: {
         reviewNo: reviewNo,
         memberNo: memberNo,
         reporter: loginMemberNo},
       success : result => {
-          console.log(result);
+          if(result==0){
+            alert("이미 신고한 리뷰입니다.");
+            return;
+          }else{
           alert("리뷰를 신고 했습니다.");
-          location.reload();
+          location.reload();}
       },
       error : error => {
         console.log(error);
+        alert("이미 신고한 리뷰입니다");
+        location.reload();
       }
     });
 }
